Extract workspace_candidates.rs from turn.rs (parent #680)
Move the workspace candidate lookup helpers out of turn.rs into a new
src/agent/loop_run/workspace_candidates.rs sibling module:

- existing_workspace_candidate_for_role (#[cfg(test)] pub(super), legacy)
- target_path_in_scope (#[cfg(test)] pub(super))
- existing_workspace_candidate_for_role_in_scope (pub(super))

The cluster filters workspace files through
`task_workspace_scope::TaskWorkspaceScope::contains` (Issue #646) and
sorts by `scaffold_pipeline::scaffold_candidate_priority`. No facade
re-export (DR3-001).

Callers updated:
- turn.rs: re-imports `existing_workspace_candidate_for_role_in_scope`
  for production callers; the two test-only fns are gated `#[cfg(test)]`.
- Drops unused `super::scaffold_pipeline::scaffold_candidate_priority`
  import from turn.rs (only used by the migrated cluster).
- Drops `meaningful_workspace_files` from turn.rs's workspace_walk import
  (only the migrated cluster called it).

turn.rs: 26,726 -> 26,646 (-80 LOC)
Cumulative from 39,489: -12,843 LOC (-32.5%)

Verification:
- cargo fmt --check
- cargo clippy --lib --all-targets -- -D warnings
- cargo test --lib (3,114 passed)
